(not-yet-working) remove middleware
