+++ frontend/package.json
@@ -116,8 +116,7 @@
     "tailwindcss": "^3.4.10",
     "tailwindcss-animate": "^1.0.7",
     "typescript": "^5.5.4",
-    "vite": "^5.3.5",
-    "vite-env-only": "^2.4.1",
+    "vite": "^5.4.2",
     "vite-tsconfig-paths": "^5.0.1",
     "vitest": "^2.0.5"
   },
